BaseAuthResource class
